+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "./App.css";
 import { MainStoreContext, mainStore } from "./stores/MainStore";
-import { GameSetup } from "./components/GameSetup";
+import { GameSetup } from "./components/game-setup/GameSetup";
 
 function App() {
   return (
