+++ src/components/game/GameField.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { observer } from "mobx-react";
-import { useMainStore } from "../hooks/useMainStore";
+import { useMainStore } from "../../hooks/useMainStore";
 
-import { Card } from "./cards/Card";
+import { Card } from "../cards/Card";
 import styles from "./GameField.module.css";
 
 export const GameField = observer(() => {
